requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-google-genai
langchain-chroma
langchain-text-splitters
chromadb
pypdf
python-docx
python-pptx
pandas
openpyxl
python-dotenv
sentence-transformers
python-multipart
